Complete overhaul of the Layer-System
changed the static system to a dynamic system because when two nodes
should be connected as decided by the inherited parental connection, but
it so happens that nodeA and nodeB are both in different layers in the
parents, there was the chance that these two nodes were expressed in the
child within the same layer and hence the connection between them
(vertical) would be invalid. Another occurence was that the Out-Node of
the connection was not in the layer it should be and hence the
connection couldnt be drawn.
Now, in every new generation all relevant (hidden) nodes are placed in
layer[1] by default and whenever a connection specifies that the nodes
need to be in different layers, the Out-Node will be moved one layer to
the right. By doing this, the network's structure (phenotype) will
appear as tight as wanted and neccessary + will fix the two described
bugs.