Major bug fix in inheritance
Although obvious in hindsight, the childrens' connections always pointed
to the parents' corresponding connection.
This caused primarily the following problem:
With 2 boards and 1 connection that are identical among the 2 boards,
when one of the boards had an AddNode mutation, the connection of the
unaffected board would also be changed (set false), since they both
referenced the same connection.
Now, when children inherit a gene, the gene is created solely for them
with the parameters of the parent's gene