Important fix and changes
1) An issue became apparent after implementing the elimination system.
When the elimination percentile determined that only one snake will be
the parent of the x new children and a mutation occured, the mutation
would occur in every child. This could have been the reason or part of
the reason why the networks topology grew very fast in short time.

2) The fitness function has been altered to reward snakes that move
   towards fruits and punish them when they move away. On top of that
they are rewarded for eating the fruit. Previously they were rewarded
for eating fruits and the amount of moves they made together with a
little punishment for snakes that focus on not crashing but not on
eating fruits. This punishment decreased for each fruit they ate along
the run. However, this fitness function caused the snakes to find a loop
hole in which they would only turn in circles to gain a massive fitness.