Major concept update + information to the state
Regarding the previous commit: All relevant code regarding evaluation
and processing of the species/snakes has been moved from the GameManager
to the Evolution class

The new concept regarding evolution is to evaluate the snakes for x
amount of runs until they engage in sexual reproduction and get
speciated. This is a very important change due to a common problem in
which the snake wants to move in a direction that it cant, e.g. moving
up when facing down. this only concerns the start of a run, i.e. the
first move the snake would make. Hence, we had a roughly 25% chance per
snake to have it not move because its first step would be illegal.
With the change, the snake gets an additional x-1 chances to prove its
abilites.
On top of that, the hidden nodes and connections are put under more
stress and cant benefit from luck, but also dont suffer from unluck.
As of now, the snakes will have 10 runs until they will sexually
reproduce and get speciated.

The program is quite stable now, though i somehow get an infinite loop
deep into runs, e.g. run 500+ which i cant allocate.
There are still some evolutionary concepts missing, e.g. elitism and a
concept in which only the top x species are allowed to reproduce if all
species' didnt improve (maxfitness) for y generations.
Speciation is working fine now, i see different topologies between
species' and that probably arised due to excluding weight as a
classifying factor, which overwhelmed the structural difference.
I might need to tweek some mutation factors though, because it
frequently happens that many hidden layers appear before there is a
substantial amount of connections.