Save&Load and Elitism
It's now possible to save networks and load them back in later.
As of now, every 30 runs a random network from the best performing
species is selected and the connections and nodes are saved into a json
file. The path is not set to be ready for release yet.

Elitism has also been implemented. When a species has at least 5
members, the best performer will be put into the next generation without
any changes.