A few changes and implementations
The visualization of the snakes' networks can now be toggled off for
faster runs, especially with big networks

The visualized nodes did not represent their real-time value, now they
do

The camera now switches to another snake when the current one has
concluded its run

Until now, there was no bias node involved in the program.
After a lot of research it seems not too necessary to have a bias node
as opposed to in other neural network models, but having different
offsets for activation hopefully diversifies the outcome of the snakes'
behavior.
There is now one bias node in the input layer which can connect to any
node that is no input node. The sigmoid function then adds the weight of
the bias node's connection on top of the rest of the inputs