Bug fixes and implementations
I tried a 3 output system (straight,left,right) which I
decided against since it would require an additional 4 input nodes to
let the snake know its direction. This however revealed a very
unpleasant bug. Normally, the snakes get their visions updated before
their first move, but due to inherting a bool from the parent which
already was set to true, not false, this vision update was ignored and
therefore the snakes first move was the last move the parent made/wanted
to make. This is solved now.

An even more important problem was that snakes were acting upon stimuli
that were not active. The reason was basically that if a snake wouldnt
make a decision to move, its former move decision was acted upon, which
makes sense if the snake would continue moving in a direction, but with
the 3 output system the snakes would therefore often run circles on end.

There is now an option to start with x% connections for all snakes with
different constellations.

Another big problem was in the feedforward function. Mistakenly, I made
the function ignore inputs from binary nodes if they were flagged with
false. However, the activation function I use, sigmoid, turns a 0 input
into 0.5 and this wouldnt happen when 0 inputs were considered null
input. This is now fixed.

Weights are now initialized with the Xavier weight initialization
method. This narrowed the range of possible weights significantly and
gave a reasonable range for weights. Additionally, since weight
distribution is now on a smaller range, I can take the average weight
difference in consideration again for speciation withhout weights being
the driving factor for speciation.