QoL adjustments and performance related changes
-Addition to last push: If a snake decides to make a 180 turn (illegal),
the snake will continue moving along the previous direction, essentially
ignoring the output decision.

-Board size can now be customized and the visuals (network,
informations) scale with the boards.

-Boards layout is now determined by the square root of the total number
of boards instead of manual input. This reduces manual inputs and helps
with creating a square-like layout for the boards.

-Previously, after each run, boards were destroyed and then
reinstantiated. Since that was performance heavy and unnecessary due to
the static role of the boards, the boards will only be instantiated once
now.

-Instead of automatically saving the best network every x runs, there is
now a save button for manual saving. Running the game in the background
and having multiple saves in the queue caused huge performance problems.